8th committ.
The GUI is working mostly. The combobox where you select the RFID does not refresh as it should and some code around that does not work as it should. Button 1 and 2 are connected to the refresh. I thought I would just remove that or fix before we would turn this thing in.
Ignore all the serial code that I have done.
It does not work and is just a mix and match of different things.